requirements.txt
flask
supabase-py
python-dateutil
uuid
time
subprocess
platform